Extract CreatesTestResponse trait for Pest compatibility
Move createTestResponse override into a reusable trait so Pest users
can mix it into their own TestCase without extending ours.
